+++ lib/findFunctionParameters.js
@@ -1,15 +1,9 @@
 const chalk = require('chalk');
-const acorn = require('acorn');
 const fs = require('fs');
 const { promisify } = require('util');
 const readFile = promisify(fs.readFile);
-const find = require('./utils/find');
 const { filter, isATypeOfFunction } = require('./utils/astParse');
-
-const findJavascriptFiles = async cwd => {
-	const result = await find('*.js', cwd);
-	return result.split('\n').filter(r => r);
-};
+const { fileToAST, findJavascriptFiles } = require('./utils/helpers');
 
 module.exports = () => {
 	const cwd = process.cwd();
@@ -38,12 +32,4 @@ module.exports = () => {
 	});
 };
 
-function fileToAST (contents) {
-	return Promise.resolve(
-		acorn.parse(contents, {
-			ecmaVersion: 8,
-			sourceType: 'module',
-			locations: true
-		}
-	));
-}
+
