+++ lib/findLongFunctions.js
@@ -1,26 +1,17 @@
 const chalk = require('chalk');
-const acorn = require('acorn');
 const fs = require('fs');
 const { promisify } = require('util');
 const readFile = promisify(fs.readFile);
-const find = require('./utils/find');
 const { filter, isATypeOfFunction } = require('./utils/astParse');
-
-const findJavascriptFiles = async cwd => {
-	const result = await find('*.js', cwd);
-	return result.split('\n').filter(r => r);
-};
+const { fileToAST, findJavascriptFiles } = require('./utils/helpers');
 
 module.exports = () => {
 	const cwd = process.cwd();
 	findJavascriptFiles(cwd).then(files => {
 		return files.forEach(file => {
-			return readFile(file, 'utf8').then(contents => {
-				const ast = acorn.parse(contents, {
-					ecmaVersion: 8,
-					sourceType: 'module',
-					locations: true
-				});
+			return readFile(file, 'utf8')
+			.then(fileToAST)
+			.then(ast => {
 				const functions = filter(ast, isATypeOfFunction({assignVariableNames: true}))
 					.map(node => {
 						const startLine = node.loc.start.line;
