+++ lib/utils/helpers.js
@@ -1,0 +1,19 @@
+const acorn = require('acorn');
+const find = require('./utils/find');
+
+function fileToAST (contents) {
+	return Promise.resolve(
+		acorn.parse(contents, {
+			ecmaVersion: 8,
+			sourceType: 'module',
+			locations: true
+		}
+	));
+}
+
+async function findJavascriptFiles (cwd) {
+	const result = await find('*.js', cwd);
+	return result.split('\n').filter(r => r);
+}
+
+module.exports = {fileToAST, findJavascriptFiles};
